Basic place bet flow

+++ src/main/java/io/gnupinguin/sporty/interview/rest/BetResource.java
@@ -1,6 +1,8 @@
 package io.gnupinguin.sporty.interview.rest;
 
+import java.math.BigDecimal;
+
 public record BetResource (
     long userId,
     long jackpotId,
-    long amount) { }
+    BigDecimal amount) { }
